refactor integaration tests
