--- test/integration/system_cpu_reader_mock_test.go
@@ -1,15 +1,0 @@
-package integration_test
-
-type SystemCPUReaderMock struct {
-	systemUsage uint64
-}
-
-func NewSystemCPUReaderMock(systemUsage uint64) *SystemCPUReaderMock {
-	return &SystemCPUReaderMock{
-		systemUsage: systemUsage,
-	}
-}
-
-func (s *SystemCPUReaderMock) ReadUsage() (uint64, error) {
-	return s.systemUsage, nil
-}
